Initial versioned release v0.0.1
Admittedly, a large commit. Included in this release and this repo are
the code for the workflow manager (often referred to as "server",
although I want to change this), and the code for an agent. Either are
meant to be replicated and still coordinate. We have tested this with
agents only.

Each of the agent and manager use the Django ORM for both the "AMQP"
process, the primary operational process for the system. Django as a
whole comes along (sharing a database) as a prepackaged administrative
interface. It doesn't do anything beyond that, and most if not all
business logic is kept out of models.  A dynamic front end is imagined
for this, but forthcoming, if ever.

Protobuf files are included and python code is generated from them as
per the README. The remainder of setup is also in the README, and is
primarily docker based. License in LICENSE.

Of note: We run django in development mode. We'll change this when we
separate out that code, or decide against doing so. Do not run in
production (yet).